♻️ Moved plot_concentrations to its own module with the same name
