+++ pyglotaran_extras/plotting/plot_overview.py
@@ -7,11 +7,11 @@ import matplotlib.pyplot as plt
 import xarray as xr
 
 from pyglotaran_extras.io.load_data import load_data
+from pyglotaran_extras.plotting.plot_concentrations import plot_concentrations
 from pyglotaran_extras.plotting.plot_residual import plot_residual
 from pyglotaran_extras.plotting.plot_spectra import plot_spectra
 from pyglotaran_extras.plotting.plot_svd import plot_svd
 from pyglotaran_extras.plotting.style import PlotStyle
-from pyglotaran_extras.plotting.utils import plot_concentrations
 
 if TYPE_CHECKING:
     from glotaran.project import Result
